Updating the README.md file to explain how to run the script.
The README.md file proposes to run the PowerShell script to provision the AzureAD applications and update the code. The manual steps are copied to a new file (Manual-Configuration-Steps.md), which is linked from the README.md for people who want to understand how to provision the applications from the portal, and how to update the configuration files.